--- src/Calendar Interface/app.json
@@ -1,13 +1,0 @@
-{
-  "pages": [
-    "pages/index/index",
-    "pages/detail/detail"
-  ],
-  "window": {
-    "backgroundTextStyle": "light",
-    "navigationBarBackgroundColor": "#617fdf",
-    "navigationBarTitleText": "事项助手",
-    "navigationBarTextStyle": "white"
-  },
-  "sitemapLocation": "sitemap.json"
-}
