--- src/Calendar Interface/app.wxss
@@ -1,68 +1,0 @@
-page {
-  font-family: 'microsoft yahei ui';
-  font-size: 28rpx;
-  background-color: white;
-  color: black;
-}
-
-/*float action*/
-.float-container {
-  position: fixed;
-  bottom: 40rpx;
-  right: 40rpx;
-  overflow: hidden;
-}
-.float-action {
-  width: 100rpx;
-  height: 100rpx;
-  border-radius: 50%;
-  box-shadow: 2px 2px 10px rgba(0, 0, 0, .5);
-  background:#617fdf;
-  z-index: 100;
-  text-align: center;
-  opacity: .5;
-  margin-left: 40rpx;
-  float: right;
-}
-.float-action:active {
-  opacity: 1
-}
-.float-action image {
-    width: 60rpx;
-    height: 60rpx;
-    margin-top: 20rpx;
-}
-
-/* .bg-normal {
-  background: #46BC62 !important;
-}
-.bg-warning {
-  background: #ECAA5B !important;
-}
-.bg-danger {
-  background: #E14848 !important;
-}
-
-.color-primary {
-  color: #46BC62 !important;
-}
-.color-warning {
-  color: #ECAA5B !important;
-}
-.color-danger {
-  color: #E14848 !important;
-}
-
-.border-normal {
-  border: 2rpx solid #46BC62 !important;
-}
-.border-warning {
-  border: 2rpx solid #ECAA5B !important;
-}
-.border-danger {
-  border: 2rpx solid #E14848 !important;
-} */
-
-.text-center {
-  text-align: center !important;
-}
