--- src/Calendar Interface/pages/index/index.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
