+++ src/miniprogram/app.wxss
@@ -17,3 +17,40 @@ button:focus{
 button::after{
   border: none;
 }
+page {
+  font-family: 'microsoft yahei ui';
+  font-size: 28rpx;
+  background-color: white;
+  color: black;
+}
+
+/*float action*/
+.float-container {
+  position: fixed;
+  bottom: 40rpx;
+  right: 40rpx;
+  overflow: hidden;
+}
+.float-action {
+  width: 100rpx;
+  height: 100rpx;
+  border-radius: 50%;
+  box-shadow: 2px 2px 10px rgba(0, 0, 0, .5);
+  background:#617fdf;
+  z-index: 100;
+  text-align: center;
+  opacity: .5;
+  margin-left: 40rpx;
+  float: right;
+}
+.float-action:active {
+  opacity: 1
+}
+.float-action image {
+    width: 60rpx;
+    height: 60rpx;
+    margin-top: 20rpx;
+}
+.text-center {
+  text-align: center !important;
+}
