+++ src/miniprogram/images/remove.png
