--- src/miniprogram/pages/calendar.json
@@ -1,3 +1,0 @@
-{
-  "usingComponents": {}
-}
