--- src/miniprogram/pages/calendar.wxml
@@ -1,2 +1,0 @@
-<!--miniprogram/pages/calendar.wxml-->
-<text>miniprogram/pages/calendar.wxml</text>
