--- src/miniprogram/pages/calendar.wxss
@@ -1,1 +1,0 @@
-/* miniprogram/pages/calendar.wxss */
