+++ src/miniprogram/pages/detail/detail.json
