+++ src/miniprogram/pages/detail/detail.wxml
