+++ src/miniprogram/pages/detail/detail.wxss
